Make reportr change the URL when selecting an interval to make reports shareable/reloadable.
This commit modifies reportr so that all controllers update the URL when changing an interval in the report. This allows sharing specific reports via the URL. Also developers can easily reload reports without having to select the interval again every time.

I also switched the require.js injection to relative paths (saves some screenspace and even bytes) and the AngularJS injection to work via the .$inject property which is also great for saving some screenspace and still works with minification.