Makes the URL to techdev portal configurable
Since we're moving the portal to be on a different host the complete URL needs to be set and should be configurable (in authorization.tpl.html).

To allow that the config system of trackr was extended a bit. The configuration file now is a require.js module (makes it easier than to load a plain JSON
file which would require a plugin). It's provided as a value that can be injected anywhere in the app. A filter to render config values to templates is added.
The config file is excluded from the dist build since it needs to be different on different environments.

During development the config file needs to be in `src/conf/`, on environments it needs to be in `conf/` (see bootstrap.prod.js).

This also allowed removing one jQuery AJAX call from jiraIssueCollector.js and the removal of the `fake` marker in the backendMock.js (the config file is now loadable
in tests, before that it broke in base.js).

Also, some unused library references where removed from test-main.js.